Add files via upload
We wanted to classify songs according to their music genre, by starting from spectograms, i.e. plots of their frequencies across time. To deal with that we started by treating spectograms as images, thus using Convolutional Neural Networks (ResNet). Then, we try to take into account the sequential meaning of these spectograms since we thought that this information would be meaningful for genre's detection. This was done by using a combination of Convolutional layers and RNN layers (we used an LSTM).